# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "description": "단기 팝업스토어 턴키 예약·무인 운영 플랫폼 API. 모든 응답은 {data, error} 봉투로 감싸진다.",
    "title": "PopupReady API",
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "로컬 개발"
    }
  ],
...
  "paths": {
...
    "/api/v1/reservation-requests": {
      "post": {
        "description": "빌더가 만든 도면을 서버에서 재검증하고 견적과 함께 예약 요청을 만든다. 그리드 범위를 벗어나거나 집기가 겹치면 400이다.",
        "operationId": "createReservationRequest",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateReservationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseReservationRequestResponse"
                }
              }
            },
            "description": "Created"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "요청 값 검증 실패"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "인증이 필요함"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "역할 또는 당사자 자격이 없음"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "본문이 지목한 공간·집기를 찾을 수 없음"
          },
          "415": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "지원하지 않는 Content-Type"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "서버 오류"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "예약 요청 생성",
        "tags": [
          "reservation"
        ]
      }
    },
...
    "/api/v1/reservation-requests/{id}/payment/confirm": {
      "post": {
        "description": "분산 락 안에서 예약 상태·기간 겹침·집기 가용량·전력 한도·금액을 재확인한 뒤 PG 승인을 호출하고, 예약을 PAID로 확정하며 분할 정산 Row를 만든다. 요청 금액은 승인 금액이 아니라 대조 대상이다.",
        "operationId": "confirmPayment",
        "parameters": [
          {
            "description": "예약 요청 ID",
            "example": 1,
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PaymentConfirmRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponsePaymentConfirmResponse"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "요청 값 검증 실패"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "인증이 필요함"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "역할 또는 당사자 자격이 없음"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "리소스를 찾을 수 없음"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "이미 결제됨 / 기간 겹침 (PAYMENT_ALREADY_COMPLETED, FIXTURE_UNAVAILABLE)"
          },
          "415": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "지원하지 않는 Content-Type"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "서버 오류"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "락 획득 실패 — 사용자 잘못이 아니므로 재시도를 안내한다 (LOCK_ACQUISITION_FAILED)"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "결제 승인",
        "tags": [
          "payment"
        ]
      }
    },
    "/api/v1/reservation-requests/{id}/payment/prepare": {
      "post": {
        "description": "토스 결제 위젯에 넘길 주문 정보를 발급한다. 락을 잡지 않으며 자리를 선점하지 않는다. orderId는 호출마다 새로 발급되고 재사용하지 않는다.",
        "operationId": "preparePayment",
        "parameters": [
          {
            "description": "예약 요청 ID",
            "example": 1,
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponsePaymentPrepareResponse"
                }
              }
            },
            "description": "Created"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "요청 값 검증 실패"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "인증이 필요함"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "역할 또는 당사자 자격이 없음"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "리소스를 찾을 수 없음"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "서버 오류"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "결제 준비",
        "tags": [
          "payment"
        ]
      }
    },
...
    "/api/v1/spaces": {
      "get": {
        "description": "중심 좌표 기준 반경(m) 안의 ACTIVE 공간을 찾는다. 면적·대여료·전력 필터는 모두 선택이다.",
        "operationId": "searchSpaces",
        "parameters": [
          {
            "description": "중심 위도",
            "example": 37.5445,
            "in": "query",
            "name": "lat",
            "required": true,
            "schema": {
              "type": "number",
              "format": "double",
              "maximum": 90.0,
              "minimum": -90.0
            }
          },
          {
            "description": "중심 경도",
            "example": 127.0557,
            "in": "query",
            "name": "lng",
            "required": true,
            "schema": {
              "type": "number",
              "format": "double",
              "maximum": 180.0,
              "minimum": -180.0
            }
          },
          {
            "description": "검색 반경(m). 미지정 시 1000",
            "example": 1500,
            "in": "query",
            "name": "radius",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1000,
              "exclusiveMinimum": 0,
              "maximum": 50000
            }
          },
          {
            "description": "최소 실면적(㎡)",
            "example": 50,
            "in": "query",
            "name": "minArea",
            "required": false,
            "schema": {
              "type": "number",
              "format": "double",
              "minimum": 0
            }
          },
          {
            "description": "일일 대여료 상한(원)",
            "example": 600000,
            "in": "query",
            "name": "maxRent",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 0
            }
          },
          {
            "description": "최소 허용 전력(W)",
            "example": 3000,
            "in": "query",
            "name": "minPower",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseListSpaceSummaryResponse"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "요청 값 검증 실패"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "서버 오류"
          }
        },
        "summary": "반경 공실 검색",
        "tags": [
          "space"
        ]
      }
    },
...
    "/api/v1/spaces/{spaceId}/fixture-availability": {
      "get": {
        "description": "해당 기간에 배치 가능한 집기 수량을 집기별로 돌려준다. 수량은 질의 기간 중 가장 많이 잡힌 날 기준이며, 다른 공간의 예약이 잡아간 수량도 함께 차감된다.",
        "operationId": "getFixtureAvailability",
        "parameters": [
          {
            "description": "대상 공간 ID",
            "example": 1,
            "in": "path",
            "name": "spaceId",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "description": "사용 시작일",
            "example": "2026-09-01",
            "in": "query",
            "name": "startDate",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "description": "사용 종료일",
            "example": "2026-09-14",
            "in": "query",
            "name": "endDate",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseListFixtureAvailabilityResponse"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "요청 값 검증 실패"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "인증이 필요함"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "리소스를 찾을 수 없음"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "서버 오류"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "날짜별 집기 가용 수량",
        "tags": [
          "reservation"
        ]
      }
    }
  },
  "components": {
    "schemas": {
      "ApiError": {
        "type": "object",
        "description": "에러 상세. 성공 응답에서는 null이다.",
        "properties": {
          "code": {
            "type": "string",
            "description": "에러 코드 상수. 클라이언트는 이 값으로 분기한다",
            "enum": [
              "VALIDATION_FAILED",
              "UNAUTHORIZED",
              "FORBIDDEN",
              "NOT_FOUND",
              "METHOD_NOT_ALLOWED",
              "UNSUPPORTED_MEDIA_TYPE",
              "INTERNAL_ERROR",
              "EMAIL_ALREADY_EXISTS",
              "INVALID_CREDENTIALS",
              "SPACE_NOT_FOUND",
              "FIXTURE_NOT_FOUND",
              "RESERVATION_REQUEST_NOT_FOUND",
              "LAYOUT_OUT_OF_BOUNDS",
              "LAYOUT_OVERLAP",
              "FIXTURE_STOCK_EXCEEDED",
              "FIXTURE_UNAVAILABLE",
              "POWER_LIMIT_EXCEEDED",
              "CONTRACT_NOT_FOUND",
              "NOT_CONTRACT_PARTY",
              "CONTRACT_ALREADY_EXISTS",
              "CONTRACT_ALREADY_SIGNED",
              "CONTRACT_INTEGRITY_VIOLATION",
              "REFRESH_TOKEN_INVALID",
              "PAYMENT_ALREADY_COMPLETED",
              "PAYMENT_AMOUNT_MISMATCH",
              "LOCK_ACQUISITION_FAILED",
              "DOOR_NOT_YET_OPENABLE"
            ],
            "example": "VALIDATION_FAILED"
          },
          "message": {
            "type": "string",
            "description": "사람이 읽는 설명. 분기 조건으로 쓰지 말 것",
            "example": "email은 필수입니다"
          }
        },
        "required": [
          "code",
          "message"
        ]
      },
      "ApiErrorResponse": {
        "type": "object",
        "description": "실패 응답 봉투. data는 항상 null이고 error에 코드·메시지가 담긴다.",
        "properties": {
          "data": {
            "description": "실패 시 항상 null"
          },
          "error": {
            "$ref": "#/components/schemas/ApiError"
          }
        },
        "required": [
          "data",
          "error"
        ]
      },
...
      "ApiResponseListFixtureAvailabilityResponse": {
        "type": "object",
        "description": "공통 응답 봉투",
        "properties": {
          "data": {
            "type": [
              "array",
              "null"
            ],
            "description": "성공 시 페이로드. 실패 시 null",
            "items": {
              "$ref": "#/components/schemas/FixtureAvailabilityResponse"
            }
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ApiError"
              },
              {
                "type": "null"
              }
            ],
            "description": "실패 시 에러 상세. 성공 시 null"
          }
        },
        "required": [
          "data",
          "error"
        ]
      },
...
      "ApiResponseListSpaceSummaryResponse": {
        "type": "object",
        "description": "공통 응답 봉투",
        "properties": {
          "data": {
            "type": [
              "array",
              "null"
            ],
            "description": "성공 시 페이로드. 실패 시 null",
            "items": {
              "$ref": "#/components/schemas/SpaceSummaryResponse"
            }
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ApiError"
              },
              {
                "type": "null"
              }
            ],
            "description": "실패 시 에러 상세. 성공 시 null"
          }
        },
        "required": [
          "data",
          "error"
        ]
      },
      "ApiResponsePaymentConfirmResponse": {
        "type": "object",
        "description": "공통 응답 봉투",
        "properties": {
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PaymentConfirmResponse"
              },
              {
                "type": "null"
              }
            ],
            "description": "성공 시 페이로드. 실패 시 null"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ApiError"
              },
              {
                "type": "null"
              }
            ],
            "description": "실패 시 에러 상세. 성공 시 null"
          }
        },
        "required": [
          "data",
          "error"
        ]
      },
      "ApiResponsePaymentPrepareResponse": {
        "type": "object",
        "description": "공통 응답 봉투",
        "properties": {
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PaymentPrepareResponse"
              },
              {
                "type": "null"
              }
            ],
            "description": "성공 시 페이로드. 실패 시 null"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ApiError"
              },
              {
                "type": "null"
              }
            ],
            "description": "실패 시 에러 상세. 성공 시 null"
          }
        },
        "required": [
          "data",
          "error"
        ]
      },
      "ApiResponseReservationRequestResponse": {
        "type": "object",
        "description": "공통 응답 봉투",
        "properties": {
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ReservationRequestResponse"
              },
              {
                "type": "null"
              }
            ],
            "description": "성공 시 페이로드. 실패 시 null"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ApiError"
              },
              {
                "type": "null"
              }
            ],
            "description": "실패 시 에러 상세. 성공 시 null"
          }
        },
        "required": [
          "data",
          "error"
        ]
      },
...
      "CreateReservationRequest": {
        "type": "object",
        "description": "예약 요청 생성",
        "properties": {
          "endDate": {
            "type": "string",
            "format": "date",
            "description": "사용 종료일(yyyy-MM-dd)",
            "example": "2026-09-14"
          },
          "layout": {
            "$ref": "#/components/schemas/LayoutDto",
            "description": "빌더가 만든 도면"
          },
          "spaceId": {
            "type": "integer",
            "format": "int64",
            "description": "대상 공간 ID",
            "example": 1
          },
          "startDate": {
            "type": "string",
            "format": "date",
            "description": "사용 시작일(yyyy-MM-dd)",
            "example": "2026-09-01"
          }
        },
        "required": [
          "endDate",
          "layout",
          "spaceId",
          "startDate"
        ]
      },
...
      "LayoutDto": {
        "type": "object",
        "description": "2D 도면 레이아웃",
        "properties": {
          "cellSizeMm": {
            "type": "integer",
            "format": "int32",
            "description": "그리드 한 칸의 실제 크기(mm)",
            "example": 500,
            "minimum": 0
          },
          "gridCols": {
            "type": "integer",
            "format": "int32",
            "description": "그리드 가로 칸 수",
            "example": 20,
            "minimum": 0
          },
          "gridRows": {
            "type": "integer",
            "format": "int32",
            "description": "그리드 세로 칸 수",
            "example": 12,
            "minimum": 0
          },
          "items": {
            "type": "array",
            "description": "배치된 집기 목록. 빈 배열도 허용한다",
            "items": {
              "$ref": "#/components/schemas/LayoutItemDto"
            }
          }
        },
        "required": [
          "cellSizeMm",
          "gridCols",
          "gridRows",
          "items"
        ]
      },
...
      "LocationDto": {
        "type": "object",
        "description": "위경도 좌표(WGS84)",
        "properties": {
          "lat": {
            "type": "number",
            "format": "double",
            "description": "위도",
            "example": 37.5445
          },
          "lng": {
            "type": "number",
            "format": "double",
            "description": "경도",
            "example": 127.0557
          }
        },
        "required": [
          "lat",
          "lng"
        ]
      },
...
      "PaymentConfirmRequest": {
        "type": "object",
        "description": "결제 승인 요청",
        "properties": {
          "amount": {
            "type": "integer",
            "format": "int64",
            "description": "클라이언트가 본 금액(원). 서버 견적과 대조만 한다",
            "example": 6930000,
            "minimum": 0
          },
          "orderId": {
            "type": "string",
            "description": "결제 준비 때 받은 주문 ID",
            "minLength": 1
          },
          "paymentKey": {
            "type": "string",
            "description": "토스가 발급한 결제 키",
            "minLength": 1
          }
        },
        "required": [
          "orderId",
          "paymentKey"
        ]
      },
      "PaymentConfirmResponse": {
        "type": "object",
        "description": "결제 승인 결과",
        "properties": {
          "amount": {
            "type": "integer",
            "format": "int64",
            "description": "승인된 금액(원)",
            "example": 6930000
          },
          "approvedAt": {
            "type": "string",
            "format": "date-time",
            "description": "PG 승인 시각(UTC)"
          },
          "orderId": {
            "type": "string",
            "description": "토스 주문 ID"
          },
          "reservationRequestId": {
            "type": "integer",
            "format": "int64",
            "description": "예약 요청 ID",
            "example": 1
          },
          "reservationStatus": {
            "type": "string",
            "description": "확정된 예약 상태",
            "enum": [
              "DRAFT",
              "CONTRACT_PENDING",
              "CONTRACT_SIGNED"
            ]
          },
          "settlements": {
            "type": "array",
            "description": "이 결제가 만든 분할 정산 Row",
            "items": {
              "$ref": "#/components/schemas/SettlementResponse"
            }
          },
          "status": {
            "type": "string",
            "description": "결제 상태",
            "enum": [
              "READY",
              "PAID",
              "FAILED",
              "CANCELLED",
              "UNKNOWN"
            ]
          }
        },
        "required": [
          "amount",
          "approvedAt",
          "orderId",
          "reservationRequestId",
          "reservationStatus",
          "settlements",
          "status"
        ]
      },
      "PaymentPrepareResponse": {
        "type": "object",
        "description": "결제 준비 결과",
        "properties": {
          "amount": {
            "type": "integer",
            "format": "int64",
            "description": "결제할 금액(원). 견적 스냅샷의 합계다",
            "example": 6930000
          },
          "orderId": {
            "type": "string",
            "description": "토스 주문 ID. 호출마다 새로 발급되며 재사용하지 않는다"
          },
          "orderName": {
            "type": "string",
            "description": "위젯에 표시할 주문명",
            "example": "성수 팝업 스페이스 14일 대여"
          }
        },
        "required": [
          "amount",
          "orderId",
          "orderName"
        ]
      },
...
      "ReservationRequestResponse": {
        "type": "object",
        "description": "예약 요청",
        "properties": {
          "brandUserId": {
            "type": "integer",
            "format": "int64",
            "description": "요청한 브랜드 사용자 ID",
            "example": 1
          },
          "endDate": {
            "type": "string",
            "format": "date",
            "description": "사용 종료일",
            "example": "2026-09-14"
          },
          "estimate": {
            "$ref": "#/components/schemas/EstimateResponse",
            "description": "견적 내역"
          },
          "id": {
            "type": "integer",
            "format": "int64",
            "description": "예약 요청 ID",
            "example": 1
          },
          "layout": {
            "$ref": "#/components/schemas/LayoutDto",
            "description": "서버가 재검증한 도면. 요청한 내용이 그대로 되돌아온다"
          },
          "spaceId": {
            "type": "integer",
            "format": "int64",
            "description": "대상 공간 ID",
            "example": 1
          },
          "startDate": {
            "type": "string",
            "format": "date",
            "description": "사용 시작일",
            "example": "2026-09-01"
          },
          "status": {
            "type": "string",
            "description": "예약 요청 상태",
            "enum": [
              "DRAFT",
              "CONTRACT_PENDING",
              "CONTRACT_SIGNED"
            ]
          }
        },
        "required": [
          "brandUserId",
          "endDate",
          "estimate",
          "id",
          "layout",
          "spaceId",
          "startDate",
          "status"
        ]
      },
...
      "SpaceSummaryResponse": {
        "type": "object",
        "description": "공간 요약(지도 마커·요약 카드용)",
        "properties": {
          "address": {
            "type": "string",
            "description": "주소",
            "example": "서울 성동구 연무장길 45"
          },
          "dailyRent": {
            "type": "integer",
            "format": "int64",
            "description": "일일 대여료(원)",
            "example": 450000
          },
          "floorAreaM2": {
            "type": "number",
            "format": "double",
            "description": "실면적(㎡)",
            "example": 82.5
          },
          "id": {
            "type": "integer",
            "format": "int64",
            "description": "공간 ID",
            "example": 1
          },
          "location": {
            "$ref": "#/components/schemas/LocationDto",
            "description": "위치 좌표"
          },
          "maxPowerWatt": {
            "type": "integer",
            "format": "int32",
            "description": "허용 전력(W)",
            "example": 5000
          },
          "name": {
            "type": "string",
            "description": "공간 이름",
            "example": "성수 연무장길 팝업 1층"
          }
        },
        "required": [
          "address",
          "dailyRent",
          "floorAreaM2",
          "id",
          "location",
          "maxPowerWatt",
          "name"
        ]
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "bearerFormat": "JWT",
        "description": "로그인 응답의 accessToken을 Bearer로 전달",
        "scheme": "bearer",
        "type": "http"
      }
    }
  }
}